+++ .github/pull_request_template.md
@@ -7,3 +7,7 @@
 ## 테스트 결과
 
 ## 리뷰 요청 사항
+
+## Check list
+- [ ] 테스트 통과
+- [ ] 문서 업데이트
