+++ package.json
@@ -13,7 +13,8 @@
     "next": "13.5.4",
     "react": "^18",
     "react-dom": "^18",
-    "sass": "^1.68.0"
+    "sass": "^1.68.0",
+    "styled-components": "^6.0.8"
   },
   "devDependencies": {
     "@types/node": "^20",
