+++ src/styles/app/home.scss
@@ -1,0 +1,39 @@
+@import '@/styles/globals';
+
+.home {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+
+  height: 100%;
+  width: 100%;
+
+  img {
+    width: 20rem;
+  }
+
+  .home__text {
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    padding-bottom: 10rem;
+    h1 {
+      font-size: 1.5rem;
+      padding-bottom: 0.5rem;
+    }
+
+    p {
+      padding-top: 0.5rem;
+    }
+  }
+
+  .home__login-text {
+    padding-top: 2rem;
+    display: inline-block;
+
+    .login {
+      color: $sub-color;
+      padding-left: 0.5rem;
+    }
+  }
+}
