+++ src/styles/app/login.scss
@@ -1,0 +1,43 @@
+@import '@/styles/globals';
+
+.sub_header {
+  font-size: 1.5rem;
+  margin: 1rem;
+  margin: 1rem 1rem 6rem 1rem;
+}
+
+.login {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  height: 100%;
+  width: 100%;
+
+  .login__text {
+    input {
+      width: 16rem;
+      height: 2rem;
+      margin-top: 1rem;
+      margin-bottom: 1rem;
+    }
+    p {
+      padding-top: 0.5rem;
+    }
+  }
+
+  .login__text {
+    padding-bottom: 10rem;
+  }
+
+  .login__pw-text {
+    padding-top: 1rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    .pw {
+      color: $sub-color;
+      padding-top: 0.5rem;
+    }
+  }
+}
