+++ src/styles/components/btn.scss
@@ -1,0 +1,19 @@
+@import '@/styles/globals';
+
+button {
+  background-color: $main-color;
+  border: none;
+  color: white;
+  padding: 1rem 2rem;
+  width: 18.5rem;
+  text-align: center;
+  text-decoration: none;
+  font-size: 1rem;
+  cursor: pointer;
+  border-radius: 0.25rem;
+  box-shadow: 0.2rem 0.2rem 0.5rem rgba(0, 0, 0, 0.4);
+
+  &:hover {
+    background-color: $sub-color;
+  }
+}
